% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{take_me2github}
\alias{take_me2github}
\title{Take user to the repo in github that corresponds to the current rstudio project}
\usage{
take_me2github(user = "tpemartin")
}
\arguments{
\item{user}{}
}
\value{

}
\description{
Take user to the repo in github that corresponds to the current rstudio project
}
